requirements.txt
jupyter
plotly==5.*
colour
scipy
spherical_geometry
tqdm==4.61.2 # https://github.com/tqdm/tqdm/issues/1199
wandb
keras-tqdm
jenkspy
# commented versions are from head_motion_prediction/README.md
protobuf==3.20.*           # prevent error "TypeError: Descriptors cannot not be created directly"
keras                      # ==2.3.1
matplotlib==3.7            # ==3.1.3
numpy==1.23.5              # ==1.18.1
opencv-python==4.5.5.64    # ==4.2.0.32
pandas==1.4.4              # ==1.0.1
pyquaternion               # ==0.9.5
quaternion==0.3.1          # ==0.3.1
scikit-image               # ==0.16.2
scikit-learn               # ==0.22.1
xlrd                       # ==1.2.0
tensorflow-macos==2.8            # tensorflow-gpu==2.1.0
# See READM for cudnn      # nvidia-cudnn-cu11==8.6.0.163
